test(IDX): isolate the advance_target_version_upgrades_all_canisters_test
The `//rs/nervous_system/integration_tests:integration_tests_test_tests/advance_sns_target_version` is currently the most flaky system-test.
It flakes because the `advance_target_version_upgrades_all_canisters_test` is timing out.
So this commit isolates this test into its own bazel target and adds some logging such that we know which part is timing out.
