+++ rs/nervous_system/integration_tests/BUILD.bazel
@@ -167,8 +167,12 @@ rust_test_suite_with_extra_srcs(
     srcs = glob(
         ["tests/**/*.rs"],
         exclude = [
+            # The following tests are flaky when run as part of this test suite.
+            # So they have been moved to dedicated bazel targets below such that we can
+            # different CPU reservations and better logging.
             "tests/sns_upgrade_test_utils.rs",
             "tests/upgrade_everything_test.rs",
+            "tests/advance_target_version_upgrades_all_canisters_test.rs",
         ],
     ),
     aliases = ALIASES,
@@ -197,6 +201,22 @@ rust_test(
     proc_macro_deps = MACRO_DEPENDENCIES + MACRO_DEV_DEPENDENCIES,
     tags = [
         "cpu:6",
+    ],
+    deps = [":nervous_system_integration_tests"] + DEPENDENCIES_WITH_TEST_FEATURES + DEV_DEPENDENCIES,
+)
+
+rust_test(
+    name = "advance_target_version_upgrades_all_canisters_test",
+    srcs = [
+        "tests/advance_target_version_upgrades_all_canisters_test.rs",
+    ],
+    aliases = ALIASES,
+    data = DEV_DATA,
+    env = DEV_ENV,
+    flaky = True,
+    proc_macro_deps = MACRO_DEPENDENCIES + MACRO_DEV_DEPENDENCIES,
+    tags = [
+        "cpu:4",
     ],
     deps = [":nervous_system_integration_tests"] + DEPENDENCIES_WITH_TEST_FEATURES + DEV_DEPENDENCIES,
 )
